+++ gsstore.go
@@ -86,14 +86,9 @@ func newGSStoreContext(ctx context.Context, baseURL *url.URL, extension, compres
 		opt.apply(&conf)
 	}
 
-	common := &commonStore{
-		compressionType:           compressionType,
-		extension:                 extension,
-		overwrite:                 overwrite,
-		uncompressedReadCallback:  conf.uncompressedReadCallback,
-		compressedReadCallback:    conf.compressedReadCallback,
-		uncompressedWriteCallback: conf.uncompressedWriteCallback,
-		compressedWriteCallback:   conf.compressedWriteCallback,
+	common, err := newCommonStore(extension, compressionType, overwrite, conf, baseURL)
+	if err != nil {
+		return nil, err
 	}
 
 	return &GSStore{