+++ s3store.go
@@ -149,14 +149,9 @@ func newS3StoreContext(ctx context.Context, baseURL *url.URL, extension, compres
 		opt.apply(&conf)
 	}
 
-	common := &commonStore{
-		compressionType:           compressionType,
-		extension:                 extension,
-		overwrite:                 overwrite,
-		uncompressedReadCallback:  conf.uncompressedReadCallback,
-		compressedReadCallback:    conf.compressedReadCallback,
-		uncompressedWriteCallback: conf.uncompressedWriteCallback,
-		compressedWriteCallback:   conf.compressedWriteCallback,
+	common, err := newCommonStore(extension, compressionType, overwrite, conf, baseURL)
+	if err != nil {
+		return nil, err
 	}
 
 	s := &S3Store{